requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-simple-history==3.7.0
